refactor: Moved Searched Book State into SearchPage Component
Moved the state for handling books found in a search into the SearchPage component instead
of in the the BookApp component. It seemed strange to have it in the BookApp component, but
I do wonder about the reusability in the SearchPage component. For this application, I do
believe it does make sense though.