bug: Can No Longer Add Same Book to Multiple Shelves
In the search page anytime you changed a book's shelf it added it to the state. So
a book could exist on multiple shelves, and even the same shelf if you changed it
from one shelf then back to the other. Now it will only add if the book doesn't
currently exist on the shelf.