bug: Can No Longer Bring Back Books from the Dead!
If a book was moved to "none" then it would still live in the state. If you were to then
add the book via search you would have two copies of the book in state and they would then
both be moved to the same shelf. Now when a book is moved to "none" it will be removed from
the state and then later added back into the state if added via search.