refactor: Moved Add Book Functionality into UpdateBooks Method in BooksApp
As in the previous commit, I hated that I had to pass down the addBook method through props,
what I liked even less is that I didn't need it to pass to the list page which would have caused
some issues if it was somehow called from that route. I moved the functionality into the App
method instead and now I only need to pass down one method to either page.