bug: Search Now Accurately Reflects Search Value
In the feedback it was noted that the search actually was being a one letter when searching. For
instance, when searching for 's' it brought up satire, but then when searching for 'sh' it continued
to bring up satire when it should have brought up Shakespeare. This is now corrected.